+++ changelog.md
@@ -1,4 +1,6 @@
-# Complicated Bees 3.4.8
+# Complicated Bees 3.5.0
+This update will break some custom centrifuge recipes, if you have any!
 
-## Fixed
-* Bee textures breaking on resource pack reloads
+## Changed
+* Centrifuge recipe definition takes a generic Ingredient input, instead of hardcoded ItemStack with component data
+* Apiary can now be crafted with vanilla honeycomb as well